+++ src/main/java/com/RINEX_parser/models/SBAS/SbasRoot.java
@@ -3,11 +3,11 @@ package com.RINEX_parser.models.SBAS;
 public class SbasRoot {
 
 	// GPS time in UTC zone
-	private long GPSTime;
+	private double GPSTime;
 
 	private long weekNo;
 
-	public long getGPStime() {
+	public double getGPStime() {
 		return GPSTime;
 	}
 
@@ -15,13 +15,13 @@ public class SbasRoot {
 		return weekNo;
 	}
 
-	public void setGPStime(long GPSTime, long weekNo) {
+	public void setGPStime(double GPSTime, long weekNo) {
 		this.GPSTime = GPSTime;
 		this.weekNo = weekNo;
 
 	}
 
-	public SbasRoot(long GPSTime, long weekNo) {
+	public SbasRoot(double GPSTime, long weekNo) {
 		super();
 		this.GPSTime = GPSTime;
 		this.weekNo = weekNo;
