+++ src/main/java/com/RINEX_parser/models/Satellite.java
@@ -14,7 +14,7 @@ public class Satellite extends Observable {
 	private double satClkDrift;
 	private double[] ECI;
 	// Note this is GPS System time at time of Reception + Receiver clock offset
-	private long tRX;
+	private double tRX;
 	// time
 	private Calendar time;
 	// Elevation and Azimuth Angle
@@ -59,7 +59,7 @@ public class Satellite extends Observable {
 	}
 
 	public Satellite(int SVID, double pseudorange, double CNo, double doppler, double phase, double carrier_frequency,
-			double[] eCEF, double satClkOff, double t, long tRX, double[] satVel, double satClkDrift, double[] ECI,
+			double[] eCEF, double satClkOff, double t, double tRX, double[] satVel, double satClkDrift, double[] ECI,
 			double[] ElevAzm, Calendar time) {
 		super(SVID, pseudorange, CNo, doppler, phase, carrier_frequency);
 		ECEF = eCEF;
@@ -73,7 +73,7 @@ public class Satellite extends Observable {
 		this.ElevAzm = ElevAzm;
 	}
 
-	public Satellite(Observable satModel, double[] eCEF, double satClkOff, double t, long tRX, double[] satVel,
+	public Satellite(Observable satModel, double[] eCEF, double satClkOff, double t, double tRX, double[] satVel,
 			double satClkDrift, double[] ECI, double[] ElevAzm, Calendar time) {
 		super(satModel);
 		ECEF = eCEF;
@@ -126,11 +126,11 @@ public class Satellite extends Observable {
 		this.satClkDrift = satClkDrift;
 	}
 
-	public long gettRX() {
+	public double gettRX() {
 		return tRX;
 	}
 
-	public void settRX(long tRX) {
+	public void settRX(double tRX) {
 		this.tRX = tRX;
 	}
